requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
tqdm
matplotlib
timm